+++ README.md
@@ -1,3 +1,69 @@
-## E-Commerce Backend
+<div align="center">
+  
+  
+  
+# Piston-Adda
 
-This folder contains the Mock Backend created for E-Commerce App. It contains different API end-points and database configuration for building E-Commerce; so that you can focus on building Frontend.
+### PistonAdda
+  
+![](https://img.shields.io/badge/React-800080?style=for-the-badge&logo=react&logoColor=white)
+![](https://img.shields.io/badge/Redux--Toolkit-593d88?style=for-the-badge&logo=redux&logoColor=white)
+![](https://img.shields.io/badge/React--Router--dom-696969?style=for-the-badge&logo=react&logoColor=white)
+
+![](https://img.shields.io/badge/HTML5-E34F26?style=for-the-badge&logo=html5&logoColor=white)
+![](https://img.shields.io/badge/CSS3-1572B6?style=for-the-badge&logo=css3&logoColor=white)
+![](https://img.shields.io/badge/JavaScript-F7DF1E?style=for-the-badge&logo=javascript&logoColor=black)
+
+</div>
+
+
+
+# About
+
+These Days social media applications are part of the daily life of every individual.This special social media website is for piston heads- AKA gear heads.
+
+##### What's different in it compared to other social media?
+
+1. As compared to others its mainly made for the individuals that are geeky and nerdy for automobiles.  
+
+
+
+
+# Features
+
+1. You can signup or login in the web app
+
+2. It's a completely responsive web app . Tablet mobile or Desktop.Ready to handle everything.
+
+3. You can post emoji or text content and images as your post and comments too.
+
+4. You can comment on the user's post's
+
+5.You can reply on comments ,upvote or downvote the replies .
+
+6. You can like and bookmark a post .
+
+7.You can edit and delete a post made by you .
+
+8.This is a whole private app .
+
+9.Can follow or unfollow a user you like.
+
+10.Can update your bio name or anything same.
+# Visit the website here
+
+## [Piston-adda](https://piston-adda.vercel.app/)
+
+
+
+
+## Connect with me
+
+ [Akshat Gupta](https://github.com/Dev-AkshatGupta)
+
+
+
+
+
+
+
