+++ src/components/index.js
@@ -4,12 +4,12 @@ export * from "./Authentication/SignUpForm";
 export * from "./CustomRoutes/PrivateRoute";
 export * from "./CustomRoutes/RestrictedRoute";
 export * from "./DifferentCards/BrandCard/BrandCard";
-export * from "./DifferentCards/Cart-Product-Big/CardProductBig";
+export * from "./DifferentCards/CartProductBig/CardProductBig";
 export * from "./DifferentCards/ProfileCard/ProfileCard";
 export * from "./DifferentCards/WishlistCard/WishlistCard";
-export * from "./DifferentCards/cartsCard/HorizontalCard";
-export * from "./DifferentCards/categoriesCard/CategoriesCard";
-export * from "./DifferentCards/reviews-card/ReviewsCard";
+export * from "./DifferentCards/CartsCard/HorizontalCard";
+export * from "./DifferentCards/CategoriesCard/CategoriesCard";
+export * from "./DifferentCards/ReviewsCard/ReviewsCard";
 export * from "./EditAddressModal/EditAddressModal";
 export * from "./ProfileAddress/ProfileAddress";
 export * from "./ProfileNav/ProfileNav";
