+++ src/pages/ProductsPage/importsAndExports.jsx
@@ -2,7 +2,7 @@
 import { NavBar } from "../../components/Navigation/NavBar";
 
 import { Footer } from "../../components/Footer/Footer";
-import { BothCard } from "../../components/DifferentCards/Cart-Product-Big/CardProductBig";
+import { BothCard } from "../../components/DifferentCards/CartProductBig/CardProductBig";
 import { useFilterManger } from "../../ContextsAndReducer/FilterDataProvider";
 import { AsideFilter } from "./ProductsPageAside/AsideFilter";
 // exports
